byte -> uint8_t, int -> int32_t

+++ DroneFirmware/BinaryHelper.h
@@ -6,32 +6,32 @@
 class BinaryHelper
 {
  public:
-	 static void changeByteOrder(byte* buffer, int start, int length);
+	 static void changeByteOrder(uint8_t* buffer, int32_t start, size_t length);
 
-	 static void writeInt16(byte* buf, int offset, int16_t val);
-	 static int16_t readInt16(byte* buf, int offset);
+	 static void writeInt16(uint8_t* buf, int32_t offset, int16_t val);
+	 static int16_t readInt16(uint8_t* buf, int32_t offset);
 
-	 static void writeInt32(byte* buf, int offset, int32_t value);
-	 static int32_t readInt32(byte* buf, int offset);
+	 static void writeInt32(uint8_t* buf, int32_t offset, int32_t value);
+	 static int32_t readInt32(uint8_t* buf, int32_t offset);
 
-	 static void writeInt64(byte* buf, int offset, int64_t val);
-	 static int64_t readInt64(byte* buf, int offset);
+	 static void writeInt64(uint8_t* buf, int32_t offset, int64_t val);
+	 static int64_t readInt64(uint8_t* buf, int32_t offset);
 
 	 
-	 static void writeUint16(byte* buf, int offset, uint16_t val);
-	 static uint16_t readUint16(byte* buf, int offset);
+	 static void writeUint16(uint8_t* buf, int32_t offset, uint16_t val);
+	 static uint16_t readUint16(uint8_t* buf, int32_t offset);
 
-	 static void writeUint32(byte* buf, int offset, uint32_t value);
-	 static uint32_t readUint32(byte* buf, int offset);
+	 static void writeUint32(uint8_t* buf, int32_t offset, uint32_t value);
+	 static uint32_t readUint32(uint8_t* buf, int32_t offset);
 
-	 static void writeUint64(byte* buf, int offset, uint64_t val);
-	 static uint64_t readUint64(byte* buf, int offset);
+	 static void writeUint64(uint8_t* buf, int32_t offset, uint64_t val);
+	 static uint64_t readUint64(uint8_t* buf, int32_t offset);
 
-	 static void writeFloat(byte* buf, int offset, float val);
-	 static float readFloat(byte* buf, int offset);
+	 static void writeFloat(uint8_t* buf, int32_t offset, float val);
+	 static float readFloat(uint8_t* buf, int32_t offset);
 
-	 static void writeDouble(byte* buf, int offset, double val);
-	 static double readDouble(byte* buf, int offset);
+	 static void writeDouble(uint8_t* buf, int32_t offset, double val);
+	 static double readDouble(uint8_t* buf, int32_t offset);
 };
 
 #endif
